move the user name lookup into a PlatformUser namespace
the name select screen had the libnx account calls and the environment lookup
inline, so a menu file included <switch.h> and talked to a system service
directly. both now sit behind PlatformUser::getUserName() in
src/framework/tools, mirroring GamePaths: a namespace of free functions that
resolves one thing per platform so the caller stays platform neutral.

the screen keeps keepSupportedChars, which is tied to its own character grid and
field length rather than to the platform, and loses both preprocessor branches.
